move stakeholder sorting into hook; simplify loading logic
